meters() return now uint
git-svn-id: https://svn.code.sf.net/p/flyhigh/code/trunk@581 1728cf9f-2348-4ec0-968f-f422e243bc8e

+++ src/common/container/WayPoint.h
@@ -53,7 +53,7 @@ class WayPoint
 		uint distance(WayPoint &wp);
 		static uint distance(WayPoint &wp1, WayPoint &wp2);
 		static double arc(uint distance);
-		static double meters(double nautmil);
+		static uint meters(double nautmil);
 		
 		WayPoint& operator=(const WayPoint &wp);
